Data object and top-level module
This change introduces two separated constructs that assist in isolating
the dependency. The first is a value object that's responsible for
representing the result of the API response for the current conditions
of a given location. It is responsible for knowing how to interact with
the JSON response from the API and transforming that into more usable
values throughout the application.

The other change is a new module at Sweathr::Weather. This becomes our
public API for any interactions to ask for the weather. It sets up a
configurable client, so that different clients can be switched out if
needed. If a client is not provided, the API class is used as a default.

This module also provides methods to interact with the weather client
provided. In this case there is one, which is the current conditions.
This will make the request to the client for the current conditions, and
transform that response into the value object
Sweathr::Weather::CurrentConditions.