Test API in isolation
This provides a spec for the API that communicates directly with Weather
Underground. Rather than testing for business logic, it's concerned with
making sure that the structure of the response is in a format that's
important for the other components of the Weather module to keep up to
date.

It uses a VCR fixture to make development speed faster by not needing to
hit the external dependency every time. However, assuming that the
cassette is part of the gitignore, it's set up to pull the API Key for
the API from an environment variable, TEST_API_KEY. This env var can be
set on the CI system, which will then make a direct call over the wire
to the resource.

Having CI call out over the wire each time will make sure that the
structure of the response stays the same. If the API should ever change
such that the response is no longer in a format that this application
expects, this test will fail and provide a warning that the application
code needs to change. This comes at the cost of being tied to the API
being up and available. Because we've isolated the API into its own
component, we can have these small-scale specs that focus on the actual
interaction with the system, but use familiar ruby mechanisms for
avoiding hitting the actual resources in the rest of our tests, whether
they be fakes, mocks or stubs.