FakeWeatherClient
This creates a fake client that does not actually communicate over HTTP.
It will instead return a seeded response, or a response based on data
passed to it from the initializer. This is then implemented in the
location spec, having that no longer rely on making a call over the wire
for the real current conditions. Now these tests can be made
deterministic, and not at the whim of the API result.