Remove VCR, Webmock examples
Now that the API is isolated in a separate class, we no longer need to
use any tools outside of our regular ruby toolbox to test the Location
class. As such, we no longer need the spec that uses VCR to fixture the
API response from Weather Underground, nor webmock to test the various
cases.