Use regular stubbing for location spec
This removes the use of the fake in favor of directly stubbing the
Weather module. This removes the need to understand the complexity of
the fake and the API. Instead, it trades that for having the test know
what the shape of the data is that it expects to be returned from
`Sweathr::Weather.current_conditions`.