Testing Mode
This provides a test mode for the overall Weather module. When enabled,
it injects a fake client as the client. The fake client has the same API
surface as the API client, but also responds to a new method which can
be used to save state for the temperatures. Rather than reaching out to
the API to retrieve the temperature, the fake client refers to an
internal hash, keyed on by zip code, to find the temperature. That value
is then provided as the current condition.