Fake Service
This creates a completely fake service, by building a Sinatra app that
responds to the same API surface that the application under test is
using for its third-party dependency. When that fake service is needed,
it is then spun up using capybara_discoball.

This involves a modification to the location class itself to support
setting/accessing the endpoint URL so that capybara_discoball can route
the request to the fake service, rather than the actual endpoint.